+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Items marked with (Serenity) are improvements due to using an updated version of
 ## [Unreleased]
 
 - Add config option for a list of keywords to cause death messages to be ignored
+- Add death message from Wardens
 - Use [anyhow](https://crates.io/crates/anyhow) for error handling
 
 ## [v2.7.0] - 2022-07-27